Team Review Cleanup
Minor comment changes and decrement deleted.
