+++ src/main/java/frc/robot/commands/indexer/IndexerShootingState.java
@@ -2,29 +2,28 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands;
+package frc.robot.commands.indexer;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.Indexer;
 import frc.robot.subsystems.RobotCargoCount;
 
 /**
- * This command implements the indexer state "Receiving."
+ * This command implements the indexer state "Shooting".
  * 
  * <p>
  * Note that the initialize method is omitted. The default implementations in
  * the commmand interface is what we need, nothing for initialize.
  * </p>
  */
-public class IndexerReceivingState extends CommandBase {
-
+public class IndexerShootingState extends CommandBase {
   /**
    * Commands in a WPILib robot implementation are often singletons but not
    * always. In the case here where we are implementing a state machine, it is a
    * best practice and sometimes required. For clarity and ease of state
    * transition implementations, we will enforce it for our state commands.
    */
-  private static IndexerReceivingState INSTANCE;
+  private static IndexerShootingState INSTANCE;
 
   /** This command's required indexer subsystem. */
   private final Indexer indexer;
@@ -35,9 +34,9 @@ public class IndexerReceivingState extends CommandBase {
    * @param indexer the indexer subsystem to use.
    * @return the command instance for the indexer state machine.
    */
-  public static final synchronized IndexerReceivingState getInstance(final Indexer indexer) {
+  public static final synchronized IndexerShootingState getInstance(final Indexer indexer) {
     if (INSTANCE == null) {
-      INSTANCE = new IndexerReceivingState(indexer);
+      INSTANCE = new IndexerShootingState(indexer);
     }
     return INSTANCE;
   }
@@ -45,50 +44,47 @@ public class IndexerReceivingState extends CommandBase {
   /**
    * @param indexer the indexer subsystem to use.
    */
-  private IndexerReceivingState(final Indexer indexer) {
+  private IndexerShootingState(final Indexer indexer) {
     this.indexer = indexer;
-    addRequirements(indexer);
-  }
-
- /**
-   * {@inheritDoc}
-   * 
-   * <p>
-   * This implementation makes sure we maintain the proper receive cargo speed.
-   * </p>
-   */
-  @Override
-  public void execute() {
-    indexer.setToReceiveVelocity();
+    addRequirements(this.indexer);
   }
 
   /**
    * {@inheritDoc}
    * 
    * <p>
-   * This command ends normally when the cargo arrives at the sensor.
+   * This implementation makes sure we maintain the proper feed cargo to shooter
+   * speed.
+   * </p>
+   */
+  @Override
+  public void execute() {
+    this.indexer.setToShooterFeedVelocity();
+  }
+
+  /**
+   * {@inheritDoc}
+   * 
+   * <p>
+   * This command ends normally when the cargo is no longer visible at the sensor.
    * </p>
    */
   @Override
   public boolean isFinished() {
-    return this.indexer.isCargoAtSensor();
-    //TODO need beam break at bottom of tower, isCargoAtSensor returns value of beam break at top of tower.
+    return !this.indexer.isCargoAtSensor();
   }
 
-
- /**
+  /**
    * {@inheritDoc}
    * 
    * <p>
    * This implementation schedules the next state in the indexer state machine.
-   * RobotCargoCount is incremented to keep track of cargo in robot.
    * </p>
    */
   @Override
   public void end(boolean interrupted) {
     this.indexer.stop();
-    RobotCargoCount.getInstance().increment();
-    IndexerReadyToShootState.getInstance(this.indexer).schedule();
-    //TODO Schedule at sensor command instead (?)
+    RobotCargoCount.getInstance().decrement();
+    IndexerEmptyState.getInstance(this.indexer).schedule();
   }
 }
